test(locale): expose test helpers
