--- packages/ubuntu_desktop_installer/test/locale/locale_model_test.mocks.dart
@@ -1,67 +1,0 @@
-// Mocks generated by Mockito 5.4.0 from annotations
-// in ubuntu_desktop_installer/test/locale/locale_model_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/services/locale_service.dart' as _i2;
-import 'package:ubuntu_desktop_installer/services/sound_service.dart' as _i4;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [LocaleService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockLocaleService extends _i1.Mock implements _i2.LocaleService {
-  MockLocaleService() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i3.Future<String> getLocale() => (super.noSuchMethod(
-        Invocation.method(
-          #getLocale,
-          [],
-        ),
-        returnValue: _i3.Future<String>.value(''),
-      ) as _i3.Future<String>);
-  @override
-  _i3.Future<void> setLocale(String? locale) => (super.noSuchMethod(
-        Invocation.method(
-          #setLocale,
-          [locale],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-}
-
-/// A class which mocks [SoundService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockSoundService extends _i1.Mock implements _i4.SoundService {
-  MockSoundService() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i3.Future<void> play(String? id) => (super.noSuchMethod(
-        Invocation.method(
-          #play,
-          [id],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-}
