+++ packages/ubuntu_desktop_installer/test/locale/locale_page_test.dart
@@ -1,47 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
-import 'package:mockito/mockito.dart';
-import 'package:ubuntu_desktop_installer/pages/locale/locale_model.dart';
 import 'package:ubuntu_desktop_installer/pages/locale/locale_page.dart';
-import 'package:ubuntu_desktop_installer/services.dart';
 import 'package:ubuntu_widgets/ubuntu_widgets.dart';
 import 'package:ubuntu_wizard/utils.dart';
 import 'package:ubuntu_wizard/widgets.dart';
 import 'package:yaru_test/yaru_test.dart';
 
 import '../test_utils.dart';
-import 'locale_model_test.mocks.dart';
-import 'locale_page_test.mocks.dart';
+import 'test_locale.dart';
 
-@GenerateMocks([TelemetryService])
 void main() {
-  LocaleModel buildModel() {
-    final locale = MockLocaleService();
-    when(locale.getLocale()).thenAnswer((_) async => 'en_US.UTF-8');
-
-    return LocaleModel(
-      locale: locale,
-      sound: MockSoundService(),
-    );
-  }
-
-  Widget buildPage(LocaleModel model) {
-    registerMockService<TelemetryService>(MockTelemetryService());
-
-    return ProviderScope(
-      overrides: [
-        localeModelProvider.overrideWith((_) => model),
-      ],
-      child: const LocalePage(),
-    );
-  }
-
   testWidgets('should display a list of languages', (tester) async {
-    final model = buildModel();
-    await tester.pumpWidget(tester.buildApp((_) => buildPage(model)));
+    final model = buildLocaleModel();
+    await tester.pumpWidget(tester.buildApp((_) => buildLocalePage(model)));
     await tester.pumpAndSettle();
 
     final languageList = find.byType(ListWidget);
@@ -93,8 +65,8 @@ void main() {
   });
 
   testWidgets('key search', (tester) async {
-    final model = buildModel();
-    await tester.pumpWidget(tester.buildApp((_) => buildPage(model)));
+    final model = buildLocaleModel();
+    await tester.pumpWidget(tester.buildApp((_) => buildLocalePage(model)));
 
     final languageList = find.byType(ListWidget);
     expect(languageList, findsOneWidget);
@@ -123,8 +95,8 @@ void main() {
   });
 
   testWidgets('should continue to next page', (tester) async {
-    final model = buildModel();
-    await tester.pumpWidget(tester.buildApp((_) => buildPage(model)));
+    final model = buildLocaleModel();
+    await tester.pumpWidget(tester.buildApp((_) => buildLocalePage(model)));
 
     final nextButton = find.button(tester.ulang.nextLabel);
     expect(nextButton, findsOneWidget);
