+++ packages/ubuntu_desktop_installer/test/locale/test_locale.mocks.dart
@@ -1,13 +1,15 @@
 // Mocks generated by Mockito 5.4.0 from annotations
-// in ubuntu_desktop_installer/test/locale/locale_page_test.dart.
+// in ubuntu_desktop_installer/test/locale/test_locale.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i3;
 
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:ubuntu_desktop_installer/services/locale_service.dart' as _i2;
+import 'package:ubuntu_desktop_installer/services/sound_service.dart' as _i4;
 import 'package:ubuntu_desktop_installer/services/telemetry_service.dart'
-    as _i2;
+    as _i5;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -20,10 +22,56 @@ import 'package:ubuntu_desktop_installer/services/telemetry_service.dart'
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
+/// A class which mocks [LocaleService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockLocaleService extends _i1.Mock implements _i2.LocaleService {
+  MockLocaleService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i3.Future<String> getLocale() => (super.noSuchMethod(
+        Invocation.method(
+          #getLocale,
+          [],
+        ),
+        returnValue: _i3.Future<String>.value(''),
+      ) as _i3.Future<String>);
+  @override
+  _i3.Future<void> setLocale(String? locale) => (super.noSuchMethod(
+        Invocation.method(
+          #setLocale,
+          [locale],
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
+}
+
+/// A class which mocks [SoundService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockSoundService extends _i1.Mock implements _i4.SoundService {
+  MockSoundService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i3.Future<void> play(String? id) => (super.noSuchMethod(
+        Invocation.method(
+          #play,
+          [id],
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
+}
+
 /// A class which mocks [TelemetryService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockTelemetryService extends _i1.Mock implements _i2.TelemetryService {
+class MockTelemetryService extends _i1.Mock implements _i5.TelemetryService {
   MockTelemetryService() {
     _i1.throwOnMissingStub(this);
   }
